Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 10.507900 10.689458  8.353201 10.383863 10.250665

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 10.038700 11.049254 10.809908 10.514179  9.662497
#> 
#> $rhs
#> $rhs$b
#> [1] 0.6279517 0.9648028 0.8850209 0.7864446 0.5025506
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  8.695091 10.514344 10.816994  6.699289  9.255687

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 11.863091 14.272887 13.055869  5.963880  7.389515
#> 
#> $rhs
#> $rhs$b
#> [1] 0.68047111 0.46609624 0.92150441 0.08532729 0.58024812
#> 
#> $rhs$cf
#> [1] 10.196259  9.553134 10.919359  8.410827  9.895590
#> 
#> $rhs$d
#> [1]  0.5556107  1.5732509  0.7121701 -0.8156492 -0.8353582
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  0.2236129 -0.1039134 -1.7059261  1.7418817 -2.2270336
#> 
#> $rhs
#> $rhs$d
#> [1]  0.1118065 -0.0519567 -0.8529631  0.8709409 -1.1135168
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  -0.6815354  -4.1403149   0.3428002 -10.3894989   0.3738192
#> 
#> $rhs
#> $rhs$d
#> [1] -0.17038385 -1.03507873  0.08570004 -2.59737474  0.09345480
#> 
#> $rhs$cf3
#> [1] -0.3407677 -2.0701575  0.1714001 -5.1947495  0.1869096
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  13.252664  -2.651413   8.276131   3.073481 -12.479593
#> 
#> $rhs
#> $rhs$d
#> [1]  2.2087773 -0.4419021  1.3793551  0.5122468 -2.0799322
#> 
#> $rhs$cf3
#> [1]  4.4175547 -0.8838042  2.7587102  1.0244936 -4.1598645
#> 
#> $rhs$cf4
#> [1]  8.835109 -1.767608  5.517420  2.048987 -8.319729
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>