Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1]  8.576234 10.584135  9.347405 10.824465  9.682026

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  8.659528 10.015534 10.184513  9.303630  9.265873
#> 
#> $rhs
#> $rhs$b
#> [1] 0.1682274 0.6202293 0.6765558 0.3829282 0.3703425
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  6.012943  8.929299 13.360013 11.596569 13.151285

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  7.896074 10.202014  8.327988 11.825830  9.870134
#> 
#> $rhs
#> $rhs$b
#> [1] 0.5194280 0.4100838 0.3592899 0.4383252 0.5678063
#> 
#> $rhs$cf
#> [1] 9.713129 9.385097 9.232715 9.469821 9.858264
#> 
#> $rhs$d
#> [1] -0.605685226  0.272305795 -0.301575660  0.785336293  0.003956456
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -2.4670836 -0.9277375 -4.0680066  1.0949846  0.4997466
#> 
#> $rhs
#> $rhs$d
#> [1] -1.2335418 -0.4638687 -2.0340033  0.5474923  0.2498733
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -1.7611694 -0.4657788  0.4454216  0.4803071  1.0435335
#> 
#> $rhs
#> $rhs$d
#> [1] -0.4402924 -0.1164447  0.1113554  0.1200768  0.2608834
#> 
#> $rhs$cf3
#> [1] -0.8805847 -0.2328894  0.2227108  0.2401536  0.5217668
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -9.7546714 12.3238142  3.5898871  1.6746520 -0.4727185
#> 
#> $rhs
#> $rhs$d
#> [1] -1.62577857  2.05396904  0.59831451  0.27910866 -0.07878642
#> 
#> $rhs$cf3
#> [1] -3.2515571  4.1079381  1.1966290  0.5582173 -0.1575728
#> 
#> $rhs$cf4
#> [1] -6.5031143  8.2158761  2.3932580  1.1164346 -0.3151457
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>